+++ src/dialog_feed.h
@@ -1,2 +1,0 @@
-#pragma once
-namespace disco_dialogues { bool InstallDialogFeed(); }